fix: address consensus-review duplication findings (drifted twins + constants) (#283)
Four fixes from the duplication/divergence consensus review (report:
docs/reviews/duplication-src-gradwave.md), all of the fixed-in-one-twin class:

- solvers/davidson.py: fold the issue-#136 fp64 subspace-eigh contract into
  the shared _rr() — CheFSI/LOBPCG's complex64 draft phases called it without
  the upcast davidson_batched's inline RR carries; returns downcast to the
  caller's block dtype so the draft stays complex64.
- scf/uspp_loop.py: give the per-k davidson_gen the indefinite-S guard its
  batched twin has (cholesky_ex catch, drop-oldest recovery, jitter fallback);
  the batched comment claimed a per-k parity that did not exist.
- constants.py: EV_A3_TO_GPA / EV_A3_TO_KBAR as the single pressure-conversion
  home; seven private copies collapsed (postscf eos/elastic/stress/stress_error,
  io/output, api/elastic) including api/relax.py's drifted CODATA-2014 value.
- api/_common.py: time_reversal_ok() and build_xc() as the single homes for the
  k≡−k predicate (3 hand-kept copies) and the exact-clone XC dispatch sites
  (both elastic sites, summary._error_estimate_xc); calculator.py's private
  _XC/_SPIN_XC now alias the canonical registries so the ASE and YAML paths
  cannot drift apart.
- tests: charged-capacitor ESM regression test asserting the strained energy
  equals esm_energy for a NET-CHARGED cell (locks #281's derivative-identity
  fix; verified to fail on the pre-fix code; the neutral-cell test cannot see
  this term).

+++ src/gradwave/solvers/davidson.py
@@ -317,9 +317,16 @@ def _rr(q: torch.Tensor, hq: torch.Tensor, nw: int) -> tuple[torch.Tensor, torch
     own s built from lazy-conj views to avoid a large-nk memory spike).
     """
     s = torch.einsum("kig,kjg->kij", q.conj(), hq)
-    s = 0.5 * (s + s.conj().transpose(-1, -2))
+    # Subspace reduction ALWAYS in fp64 (issue #136), same contract as
+    # davidson_batched's inline RR: an fp32 eigh of an ill-conditioned subspace
+    # matrix in the complex64 draft phase yields inaccurate Ritz rotations.
+    # Upcast only the small (nk, m, m) matrix and the eigensolve; returns are
+    # downcast so the caller's basis and Ritz combination stay in the block
+    # dtype. (davidson_batched keeps its matmul-built s for memory, but the
+    # fp64 contract is now identical in both.)
+    s = (0.5 * (s + s.conj().transpose(-1, -2))).to(torch.complex128)
     w, u = _eigh_subspace(s)
-    return w[:, :nw].real, u[:, :, :nw]
+    return w[:, :nw].real.to(q.real.dtype), u[:, :, :nw].to(q.dtype)
 
 
 def _combine(c: torch.Tensor, block: torch.Tensor) -> torch.Tensor:
